refactor: keep passive coordination modules bounded
